Add installation information for the tarball users and from a git checkout.
We want to have an INSTALL file that contains installation instructions for
the tarball users. These instructions are necessarily package specific,
because they need to list the prerequisites or reference the DEPENDENCIES
file.

Additionally, many people who are not aware that tarballs are the primary
means for source code distribution of GNU packages check out the git repository
and then expect to see instructions for installing from that git checkout
in the INSTALL file. See e.g. <https://savannah.gnu.org/bugs/?62196>.

* Makefile.maint (gnulib-imported-files): Import INSTALL.generic from gnulib.
* INSTALL: New file, based on poke/INSTALL.
* HACKING: Renamed from README-hacking.
* Makefile.in (SOURCE_FILES): Add INSTALL, INSTALL.generic, HACKING. Remove
README-hacking.